requirements.txt
Django>=4.2,<5
djangorestframework
psycopg2-binary
requests
python-dotenv
